backend/requirements.txt
tinyfish
fastapi
uvicorn[standard]
pymongo
python-dotenv
apscheduler
httpx
agentops
pydantic-settings
